Add files via upload
Rewrote to make sure all packages are being pulled in from the vdci construction and instrument code, as they were not earlier, and the vdci construction and instrument codes was therefore not working. Moreover, I think it is better to do it like this since packages such as SVARS do not seem to be on CRAN and need to be therefore installed for new users when the code is shared.